requirements.txt
Django>=4.2
python-nmap>=0.7.1
